Resolved mask problem: player was lit if behind the desk.
Reason: the mask is performed by a Light2D, which masks AND acts as a light emitter.
Solution: set the color of the light emitted by the mask to be the same as the canvasmodulate.
TODO: link these colors together to that if CanvaModulate changes, the mask light color follows the same color.